+++ CHANGELOG.md
@@ -12,3 +12,4 @@ Here is a list of changes for each of the releases.
 
 - Formatter no longer forces BOM into files since `MASM` doesn't recognize it
 - Line breaks `LF` and `CRLF` are preserved instead of forcing `CRLF`
+- Added `--linebreaks` option to force specific line breaks